# apps/mas-ops-api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClientSummaryResponse": {
        "description": "Client summary read-model response.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "critical_asset_count": {
            "title": "Critical Asset Count",
            "type": "integer"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "open_alert_count": {
            "title": "Open Alert Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "fabric_id",
          "name",
          "open_alert_count",
          "critical_asset_count",
          "updated_at"
        ],
        "title": "ClientSummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentResponse": {
        "description": "Incident detail response.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "client_id",
          "fabric_id",
          "state",
          "severity",
          "summary",
          "opened_at",
          "updated_at"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MAS Ops API",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clients/{client_id}": {
      "get": {
        "description": "Fetch one authorized client summary.",
        "operationId": "get_client_clients__client_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Client",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/incidents/{incident_id}": {
      "get": {
        "description": "Fetch one incident detail.",
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/streams/chat/{chat_session_id}": {
      "get": {
        "description": "Stream updates for one authorized chat session.",
        "operationId": "stream_chat_streams_chat__chat_session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_session_id",
            "required": true,
            "schema": {
              "title": "Chat Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "replay_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Replay Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Chat",
        "tags": [
          "streams"
        ]
      }
    },
    "/streams/clients/{client_id}": {
      "get": {
        "description": "Stream updates for one authorized client.",
        "operationId": "stream_client_streams_clients__client_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "replay_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Replay Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Client",
        "tags": [
          "streams"
        ]
      }
    },
    "/streams/incidents/{incident_id}": {
      "get": {
        "description": "Stream updates for one authorized incident.",
        "operationId": "stream_incident_streams_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "replay_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Replay Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Incident",
        "tags": [
          "streams"
        ]
      }
    },
    "/streams/portfolio": {
      "get": {
        "description": "Stream authorized portfolio updates.",
        "operationId": "stream_portfolio_streams_portfolio_get",
        "parameters": [
          {
            "in": "query",
            "name": "replay_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Replay Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Portfolio",
        "tags": [
          "streams"
        ]
      }
    }
  }
}