# apps/mas-ops-ui/src/api/generated/types.ts
// Generated by apps/mas-ops-ui/scripts/generate_openapi_client.py
// Do not edit by hand.

export type ActivityEventResponse = {
  activity_id: number;
  asset_id: string | null;
  client_id: string;
  event_type: string;
  fabric_id: string;
  incident_id: string | null;
  occurred_at: string;
  payload: Record<string, unknown>;
  source_event_id: string;
  subject_id: string;
  subject_type: string;
};

export type ApprovalDecisionRequest = {
  decision: ApprovalDecisionValue;
  reason?: string | null;
};

export type ApprovalDecisionValue = "approve" | "reject";

export type ApprovalResponse = {
  action_kind: string;
  approval_id: string;
  approved_at: string | null;
  cancelled_at: string | null;
  client_id: string;
  decided_at: string | null;
  decided_by_user_id: string | null;
  decision_reason: string | null;
  executed_at: string | null;
  expired_at: string | null;
  expires_at: string;
  fabric_id: string;
  incident_id: string | null;
  payload: Record<string, unknown>;
  rejected_at: string | null;
  requested_at: string;
  requested_by_agent: string;
  risk_summary: string;
  state: ApprovalState;
  title: string;
};

export type ApprovalState = "pending" | "approved" | "rejected" | "expired" | "cancelled" | "executed";

export type AssetResponse = {
  asset_id: string;
  asset_kind: string;
  client_id: string;
  fabric_id: string;
  health_observed_at: string | null;
  health_state: string;
  hostname: string | null;
  last_alert_at: string | null;
  mgmt_address: string | null;
  model: string | null;
  site: string | null;
  tags: Array<string>;
  updated_at: string;
  vendor: string | null;
};

export type ChatMessageCreateRequest = {
  message: string;
};

export type ChatMessageResponse = {
  chat_session_id: string;
  content: string;
  created_at: string;
  message_id: string;
  role: string;
  turn_id: string | null;
};

export type ChatScope = "global" | "incident";

export type ChatSessionCreateRequest = {
  client_id?: string | null;
  fabric_id?: string | null;
  incident_id?: string | null;
  scope: ChatScope;
};

export type ChatSessionResponse = {
  chat_session_id: string;
  client_id: string | null;
  created_at: string;
  created_by_user_id: string;
  fabric_id: string | null;
  incident_id: string | null;
  messages: Array<ChatMessageResponse>;
  scope: ChatScope;
  turns: Array<ChatTurnResponse>;
};

export type ChatTurnResponse = {
  actor_user_id: string;
  approval_id: string | null;
  chat_session_id: string;
  completed_at: string | null;
  request_id: string;
  state: ChatTurnState;
  submitted_at: string;
  turn_id: string;
};

export type ChatTurnState = "running" | "waiting_for_approval" | "completed" | "failed" | "cancelled";

export type ClientSummaryResponse = {
  client_id: string;
  critical_asset_count: number;
  fabric_id: string;
  name: string;
  open_alert_count: number;
  updated_at: string;
};

export type ConfigApplyResult = {
  client_id: string;
  completed_at?: string | null;
  config_apply_run_id: string;
  desired_state_version: number;
  error_summary?: string | null;
  started_at?: string | null;
  status: ConfigApplyState;
};

export type ConfigApplyState = "pending" | "validating" | "applying" | "succeeded" | "failed" | "cancelled";

export type ConfigDesiredStateInput = {
  client_id: string;
  desired_state_version: number;
  fabric_id: string;
  inventory_sources?: Array<Record<string, JSONValueInput>>;
  notification_routes?: Array<Record<string, JSONValueInput>>;
  policy?: Record<string, JSONValueInput>;
  tenant_metadata?: Record<string, JSONValueInput>;
};

export type ConfigDesiredStateOutput = {
  client_id: string;
  desired_state_version: number;
  fabric_id: string;
  inventory_sources?: Array<Record<string, JSONValueOutput>>;
  notification_routes?: Array<Record<string, JSONValueOutput>>;
  policy?: Record<string, JSONValueOutput>;
  tenant_metadata?: Record<string, JSONValueOutput>;
};

export type ConfigValidationResult = {
  client_id: string;
  config_apply_run_id: string;
  desired_state_version: number;
  errors?: Array<string>;
  status: string;
  validated_at: string;
  warnings?: Array<string>;
};

export type HTTPValidationError = {
  detail?: Array<ValidationError>;
};

export type IncidentResponse = {
  client_id: string;
  fabric_id: string;
  incident_id: string;
  opened_at: string;
  severity: string;
  state: string;
  summary: string;
  updated_at: string;
};

export type JSONValueInput = boolean | number | string | JSONArrayInput | JSONObjectInput | null;

export interface JSONObjectInput {
  [key: string]: JSONValueInput;
}

export interface JSONArrayInput extends Array<JSONValueInput> {}

export type JSONValueOutput = boolean | number | string | JSONArrayOutput | JSONObjectOutput | null;

export interface JSONObjectOutput {
  [key: string]: JSONValueOutput;
}

export interface JSONArrayOutput extends Array<JSONValueOutput> {}

export type LoginRequest = {
  email: string;
  password: string;
};

export type SessionResponse = {
  client_ids: Array<string>;
  display_name: string;
  email: string;
  role: UserRole;
  user_id: string;
};

export type UserRole = "admin" | "operator" | "viewer";

export type ValidationError = {
  ctx?: Record<string, unknown>;
  input?: unknown;
  loc: Array<string | number>;
  msg: string;
  type: string;
};